commands: remove __init__ and factor out TestsetResult
