--- internal/commands/__init__.py
@@ -1,15 +1,0 @@
-from .gen import command_gen
-from .invoke import command_invoke
-from .clean import command_clean
-from .export import command_export
-from .make_public import command_make_public
-from .verify import command_verify
-
-__all__ = [
-    "command_gen",
-    "command_invoke",
-    "command_clean",
-    "command_export",
-    "command_make_public",
-    "command_verify",
-]
