+++ tests/test_gen.py
@@ -13,7 +13,7 @@ from internal.outcomes import (
     GenerationResult,
     ExecutionOutcome,
 )
-from internal.commands import command_clean
+from internal.commands.clean import command_clean
 from internal.commands.gen import command_gen
 from internal.steps.utils import CompilationSlot
