+++ tests/test_make_public.py
@@ -5,7 +5,9 @@ import pytest
 from internal.context import TMTContext
 from internal.formatting.terminal import TerminalFormatter
 
-from internal.commands import command_clean, command_gen, command_make_public
+from internal.commands.clean import command_clean
+from internal.commands.gen import command_gen
+from internal.commands.make_public import command_make_public
 
 
 expected_results_communication_2_proc_grader_stdio = {
